Move directoryUtils and cleanup
